requirements.txt
numpy>=2.4.1
matplotlib>=3.10.8
numba>=0.65.0
pillow>=12.1.0
scikit-image>=0.26.0
colordict>=1.2.6
